Finished adapting SFAgents to retro.
Added an evaluate method to visualize the trained model.

Adaptation is not entirely complete. For some reason after a bit of
training, all the model does is jump and block, no attacking. This needs
further training to see what the issue is, or if this is just a lack of
training. Also, a conversion between Mortal Kombat and Street Fighter.
Right now, the preprocessing in WorkerUtils should only work for
streetfighter.